Put live the large tag-map view

+++ db/migrate/20130306044555_create_photos.rb
@@ -2,8 +2,8 @@ class CreatePhotos < ActiveRecord::Migration
   def change
     create_table :photos do |t|
       t.string :photo, :limit => 250
-      t.decimal :lat
-      t.decimal :lon
+      t.float :lat
+      t.float :lon
       t.string :caption, :limit => 250
       t.integer :assignment_id
